move the psycopg2 database connection to the flask app context
This ensures that we only ever create a single connection per request,
and also removes the hack for the `db.adapt()` function. Furthermore,
we now do the caching within the get_connection() method, it no longer
matters how many times we call it. Hence, I killed that test.

The motivation for this is twofold: first of all, not having the
adapt() connection stick around should make it easier to kick the
PostgreSQL processes without also bouncing OIO REST, and second, it
should aid with running OIO REST in PyPy.* Also, it's nice and the
documented way of doing things.

* I know, no-one cares about this but me…